requirements.txt
python-vlc
requests
pypresence
PySimpleGUI